% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compFDR.R
\name{compFDR}
\alias{compFDR}
\title{Benchmark for gene set analysis methods using 24 datasets}
\usage{
compFDR(datasets = NULL, existingMethods = c("GSA", "PADOG"),
  mymethods = NULL, gs.names = NULL, gslist = "KEGG.db",
  organism = "hsa", Nmin = 3, NI = 1000, use.parallel = TRUE,
  ncr = NULL, pkgs = "GSA", expVars = NULL, dseed = NULL, Npsudo = 20,
  FDRmeth = c("BH", "Permutation"), plots = FALSE, verbose = FALSE)
}
\arguments{
\item{datasets}{A character vector with valid names of datasets to use from the PADOGsets
package. If left NULL all datasets avalibale in PADOGsets will be used.}

\item{existingMethods}{A character vector with one or more of the predefined methods
c("GSA","PADOG"). The first is used as reference method.}

\item{mymethods}{A list whose elements are valid functions implementing gene set analysis
methods. See \code{Details} and \code{Examples} for what arguments the functions have to
take in and what kind of output they need to return.}

\item{gs.names}{A character vector giving additional information about each gene set. For
instance when gene seta are pathways, the full name of the pathway would be a meaningful
gene set name.}

\item{gslist}{Either the value "KEGG.db" or a list with the gene sets. If set to "KEGG.db",
then gene sets will be made of all KEGG pathways for human since all datasets available
in PADOG are for human.}

\item{organism}{A three letter string giving the name of the organism supported by the
"KEGG.db" package.}

\item{Nmin}{The minimum size of gene sets to be included in the analysis for all methods.}

\item{NI}{Number of iterations to determine the gene set score significance p-values in
PADOG and GSA methods.}

\item{use.parallel}{If set to TRUE and multiple CPU cores are available, the parallelization
will be distributed either over the \code{NI} iterations for "PADOG", or over
datasets/methods combination otherwise.}

\item{ncr}{The number of CPU cores used when \code{use.parallel} set to TRUE. Default is
to use all CPU cores detected.}

\item{pkgs}{Character vector of packages that the \code{existingMethods} and \code{mymethods}
depend on (e.g. \code{NULL} for "PADOG", "GSA" for "GSA"). Consult the \code{.packages}
argument in \code{foreach} function from \code{foreach} package.}

\item{expVars}{Character vector of variables to export. Consult the \code{.export} argument
in \code{foreach} function from \code{foreach} package.}

\item{dseed}{Optional initial seed for random number generator (integer) used in \code{padog}.}

\item{Npsudo}{The number of permutations on phenotype (i.e. group label) to obtain null p
values and estimate fdr. Set to 0 if not interested in estimating fdr from permutation.}

\item{FDRmeth}{The method used to estimate fdr; "BH" for Benjamini & Hochberg, "Permutation"
for estimation based on observed and null p values.}

\item{plots}{If set to TRUE will plot the p values, ranks, the ranks differences w.r.t.
reference, null p values and fdr when applicable.}

\item{verbose}{This argument will be passed to PADOG and AbsmT methods. If set to TRUE
it will show the iterations performed so far (when \code{use.parallel = TRUE}, this only
works for command line R, not Rgui).}
}
\value{
A list of elements "summary","ranks","pvalues","qvalues" and "nullp" (if \code{Npsudo > 0}).
  "summary" is a data frame containing: \code{Method} is the name of the gene set analysis method;
    \code{p geomean} geometric mean of nominal p-values for the target gene sets (gene sets expected
    to be relevant); \code{p med} median of nominal p-values for the target genesets; \code{ \% p<0.05}
    is the fraction of target gene sets significant at 0.05 level (this is the sensitivity);
    \code{ \% q<0.05} is the fraction of target gene sets significant at 0.05 level after FDR correction;
    \code{rank mean} mean rank of the target gene sets; \code{rank med} median rank of the target gene sets;
    \code{p Wilcox.} p value from a Wilcoxon test paired at dataset level comparing the rank of target gene sets;
    \code{p LME} p value from a linear mixed effects (LME) model which unlike the Wilcoxon test above
      accounts for the fact that ranks for the same gene set may be correlated;
    \code{coef LME} Coefficient from the LME model giving the difference in ranks of the target gene
      sets between the current gene set analysis method and the reference method (the first method
      in the \code{existingMethods} argument.
  "ranks", "pvalues", "qvalues": a list of elements containing for each method the ranks (
    nominal p values, fdr q values) of the target gene sets in the corresponding data sets.
  "nullp": a list of elements containing for each method the null p values (a list of two
    elements "ap" for all gene sets and "tp" for target gene sets) pooled over all data sets.
}
\description{
This is a general purpose function to compare a given gene set analysis method in terms of
sensitivity, ranking and fdr against PADOG and GSA (if installed) using 24 public datasets.
}
\details{
The user defined gene set analysis method should return a list with elements "targ" containing
the analysis statistics for the target gene sets on a given data set, and "pval" containing
a matrix of null p values each column of which corresponds to the p values of all gene sets
applying the gene set analysis method on permutated phenotype (i.e. group label). The rownames
of the matrix are the gene set IDs; the number of column corresponds to argument \code{Npsudo}.
If \code{Npsudo = 0}, the "pval" element can be omitted and null p values / fdr will not be
compared between the methods. For "PADOG" and "AbsmT", two methods ("BH" and "Permutation") of
estimating fdr are available regardless of the value of \code{Npsudo} since null p values can be
accessed from the method algorithm itself. For "GSA", the "Permutation" method is only available
when \code{Npsudo > 0}.
}
\examples{
#compare a new geneset analysis method with PADOG and GSA

#define your new gene set analysis method that takes as input:
#set- the name of dataset file from the PADOGsetspackage
#mygslist - a list with the genesets
#minsize- minimum number of genes in a geneset to be considered for analysis

randomF=function(set,mygslist,minsize){
set.seed(1)
#this loads the dataset in an ExpressionSet object called x
data(list=set,package="KEGGdzPathwaysGEO")
x=get(set)

#Extract from the dataset the required info to be passed to padog
exp=experimentData(x);
dat.m=exprs(x)
ano=pData(x)
dataset= exp@name
design= notes(exp)$design
annotation= paste(x@annotation,".db",sep="")
targetGeneSets= notes(exp)$targetGeneSets


#get rid of duplicates probesets per ENTREZ ID by keeping the probeset
#with smallest p-value (computed using limma)
aT1=filteranot(esetm=dat.m,group=ano$Group,paired=(design=="Paired"),
 block=ano$Block,annotation=annotation)
#create an output dataframe for this toy method with random gene set p-values
mygslistSize=unlist(lapply(mygslist,function(x){length(intersect(aT1$ENTREZID,x))}))
res=data.frame(ID=names(mygslist),P=runif(length(mygslist)),
 Size=mygslistSize,stringsAsFactors=FALSE)
res$FDR=p.adjust(res$P,"fdr")
#drop genesets with less than minsize genes in the current dataset
res=res[res$Size>=minsize,]
#obtain null p value matrix
nullp = sapply(1:20, function(x)  runif(nrow(res)))
rownames(nullp) = res$ID
#compute ranks
res$Rank=rank(res$P)/dim(res)[1]*100
#needed to compare ranks between methods; must be the same as given
#in mymethods argument "list(myRand="
res$Method="myRand";
#needed because comparisons of ranks between methods is paired at dataset level
res$Dataset<-dataset;
#output only result for the targetGeneSets
#which are gene sets expected to be relevant in this dataset
return(list(targ = res[res$ID \%in\% targetGeneSets,], pval = nullp))
}

#run the analysis on all 24 datasets and compare the new method "myRand" with
#PADOG and GSA (if installed) (chosen as reference since is listed first in the existingMethods)
\dontrun{
  out = compFDR(datasets=NULL,existingMethods=c("GSA","PADOG"),
    mymethods=list(myRand=randomF),gslist="KEGG.db",Nmin=3,NI=1000,Npsudo=20, FDRmeth="P",
    plots=FALSE,verbose=FALSE,use.parallel=TRUE,dseed=1,pkgs=c("GSA","PADOG"))
}

#compare myRand against PADOG on 3 datasets only
#mysets=data(package="PADOGsets")$results[,"Item"]
mysets=c("GSE9348","GSE8671","GSE1297")
out = compFDR(datasets=mysets,existingMethods=c("PADOG"),
  mymethods=list(myRand=randomF),gslist="KEGG.db",Nmin=3,NI=20,Npsudo=0,
  plots=FALSE,verbose=FALSE,use.parallel=FALSE,dseed=1,pkgs=NULL)
}
\author{
Adi Laurentiu Tarca \email{atarca@med.wayne.edu},
   Zhonghui Xu \email{zhonghui.xu@gmail.com}
}
\references{
Adi L. Tarca, Sorin Draghici, Gaurav Bhatti, Roberto Romero. Down-weighting
   overlapping genes improves gene set analysis. BMC Bioinformatics, 2012.
}
\seealso{
See \code{\link{compPADOG}} for back-compatibility (i.e. no comparing fdr).
}
\keyword{methods}
\keyword{nonparametric}

