% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padog.R
\name{padog}
\alias{padog}
\title{Pathway Analysis with Down-weighting of Overlapping Genes (PADOG)}
\usage{
padog(esetm = NULL, group = NULL, paired = FALSE, block = NULL,
  gs = NULL, NI = 1000, Nmin = 3, verbose = TRUE,
  parallel = FALSE, dseed = NULL, ncr = NULL, rna_seq = FALSE,
  pdata = NULL)
}
\arguments{
\item{gs}{Named list with: \itemize{
  \item \code{gs$list}  named lists (KEGG pathway number) for each pathway, each with a character vectors of entrez ids with \code{names} attribute equal to HGNC symbols.
  \item \code{gs$names}  named character vector of KEGG pathway names with \code{names} attribute equal to KEGG pathway numbers.
}}

\item{rna_seq}{is the analysis on RNA seq data? Default is \code{FALSE}. If \code{TRUE} must supply \code{pdata}.}

\item{pdata}{data.frame with columns \code{lib.size} and \code{norm.factors} needed if \code{rna_seq} is \code{TRUE}.}
}
\description{
This is a general purpose gene set analysis method that downplays the
importance of genes that apear often accross the sets of genes analyzed.
}
\details{
The original implementation of \code{\link[PADOG]{padog}} was modified to allow two-sample
groups and eliminate defunct KEGG.db warning.
}
